com.utilities.audio 1.0.8
- Added PCMEncoder class for easier sample conversion
