+++ Utilities.Audio/Packages/com.utilities.audio/Runtime/AudioClipExtensions.cs
@@ -1,8 +1,6 @@
 // Licensed under the MIT License. See LICENSE in the project root for license information.
 
-using System;
 using UnityEngine;
-using UnityEngine.Assertions;
 
 namespace Utilities.Audio
 {
@@ -18,58 +16,8 @@ namespace Utilities.Audio
         public static byte[] EncodeToPCM(this AudioClip audioClip, PCMFormatSize size = PCMFormatSize.EightBit, bool trim = false)
         {
             var samples = new float[audioClip.samples * audioClip.channels];
-            var sampleCount = samples.Length;
             audioClip.GetData(samples, 0);
-
-            // trim data
-            var start = 0;
-            var end = sampleCount;
-
-            if (trim)
-            {
-                for (var i = 0; i < sampleCount; i++)
-                {
-                    if (samples[i] * short.MaxValue == 0)
-                    {
-                        continue;
-                    }
-
-                    start = i;
-                    break;
-                }
-
-                for (var i = sampleCount - 1; i >= 0; i--)
-                {
-                    if (samples[i] * short.MaxValue == 0)
-                    {
-                        continue;
-                    }
-
-                    end = i + 1;
-                    break;
-                }
-            }
-
-            var trimmedLength = end - start;
-            Assert.IsTrue(trimmedLength > 0);
-            Assert.IsTrue(trimmedLength <= sampleCount);
-            var sampleIndex = 0;
-            var pcmData = size switch
-            {
-                PCMFormatSize.EightBit => new byte[trimmedLength * sizeof(short)],
-                PCMFormatSize.SixteenBit => new byte[trimmedLength * sizeof(float)],
-                _ => throw new ArgumentOutOfRangeException(nameof(size), size, null)
-            };
-
-            // convert and write data
-            for (var i = start; i < end; i++)
-            {
-                var sample = (short)(samples[i] * short.MaxValue);
-                pcmData[sampleIndex++] = (byte)(sample >> 0);
-                pcmData[sampleIndex++] = (byte)(sample >> 8);
-            }
-
-            return pcmData;
+            return PCMEncoder.Encode(samples, size, trim);
         }
 
         /// <summary>
@@ -80,29 +28,7 @@ namespace Utilities.Audio
         /// <param name="size">Size of PCM sample data.</param>
         public static void DecodeFromPCM(this AudioClip audioClip, byte[] pcmData, PCMFormatSize size = PCMFormatSize.EightBit)
         {
-            var sampleCount = pcmData.Length / (sizeof(short) * (int)size);
-            var samples = new float[sampleCount];
-            var sampleIndex = 0;
-
-            switch (size)
-            {
-                case PCMFormatSize.EightBit:
-                    for (var i = 0; i < pcmData.Length; i++)
-                    {
-                        samples[sampleIndex++] = pcmData[i] / 128f - 1f;
-                    }
-                    break;
-                case PCMFormatSize.SixteenBit:
-                    for (var i = 0; i < pcmData.Length; i += 2)
-                    {
-                        var sample = (short)((pcmData[i + 1] << 8) | pcmData[i]);
-                        samples[sampleIndex++] = sample / (float)short.MaxValue;
-                    }
-                    break;
-                default:
-                    throw new ArgumentOutOfRangeException(nameof(size), size, null);
-            }
-
+            var samples = PCMEncoder.Decode(pcmData, size);
             // Set the decoded audio data directly into the existing AudioClip
             audioClip.SetData(samples, 0);
         }
