+++ Utilities.Audio/Packages/com.utilities.audio/Runtime/PCMEncoder.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 33f387444f1012443a64bbf067a215f0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
